removed decode traits, turned into functions
