+++ src/lib.rs
@@ -7,7 +7,7 @@ mod debugger;
 mod jit;
 
 pub use crate::core::{
-    ARMCore, ARMv4, ARMv4Decode, Thumbv4Decode, ARMv4Instruction, CPSR, SwiHook
+    ARMCore, ARMv4, ARMv4Instruction, CPSR, SwiHook, decode_arm_v4, decode_thumb_v4
 };
 
 pub use crate::memory::{
